Make rename closer to correct.
GridFTP assumes the rename semantics are the same as POSIX: the
operation is atomic and overwrites the destination file.

This is not true for the version of rename exposed by libhdfs - it
does not overwrite.  So, on failures, we do a non-atomic delete
plus rename.  It's not the correct semantics, but the best we can
do with this version of HDFS.